+++ packages/mobility_features/test/mobility_features_test.dart
@@ -100,163 +100,9 @@ void main() async {
       }
     });
 
-    test('Features: Single Stop', () async {
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-      Duration timeTracked = Duration(hours: 17);
-
-      /// Collect location samples (synthetic data set)
-      List<LocationSample> samples = [
-        // home from 00 to 17
-        LocationSample(pos1, jan01),
-        LocationSample(pos1, jan01.add(timeTracked)),
-      ];
-
-      /// Save samples to disk
-      await mf.saveSamples(samples);
-
-      /// Compute features
-      MobilityContext context = await mf.computeFeatures(date: jan01);
-
-      expect(context.homeStay, 1.0);
-      expect(context.stops.length, 1);
-      expect(context.moves.length, 0);
-      expect(context.places.length, 1);
-    });
-
-    test('Features: Single day, multiple locations', () async {
-      /// Clean file every time test is run
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-      mf.usePriorContexts = true;
-
-      List<LocationSample> locationSamples = [
-        // 5 hours spent at home
-        LocationSample(pos1, jan01.add(Duration(hours: 0, minutes: 0))),
-        LocationSample(pos1, jan01.add(Duration(hours: 6, minutes: 0))),
-
-        LocationSample(pos2, jan01.add(Duration(hours: 8, minutes: 0))),
-        LocationSample(pos2, jan01.add(Duration(hours: 9, minutes: 30))),
-
-        LocationSample(pos3, jan01.add(Duration(hours: 10, minutes: 0))),
-        LocationSample(pos3, jan01.add(Duration(hours: 11, minutes: 30))),
-
-        /// 1 hour spent at home
-        LocationSample(pos1, jan01.add(Duration(hours: 15, minutes: 0))),
-        LocationSample(pos1, jan01.add(Duration(hours: 16, minutes: 0))),
-
-        LocationSample(pos4, jan01.add(Duration(hours: 17, minutes: 0))),
-        LocationSample(pos4, jan01.add(Duration(hours: 18, minutes: 0))),
-
-        // 1 hour spent at home
-        LocationSample(pos1, jan01.add(Duration(hours: 20, minutes: 0))),
-        LocationSample(pos1, jan01.add(Duration(hours: 21, minutes: 0))),
-      ];
-
-      await mf.saveSamples(locationSamples);
-
-      /// Calculate and save context
-      MobilityContext context = await mf.computeFeatures(date: jan01);
-
-      Duration homeTime = Duration(hours: 8);
-      Duration timeTracked =
-          Duration(hours: locationSamples.last.datetime.hour);
-      double homeStayTruth =
-          homeTime.inMilliseconds / timeTracked.inMilliseconds;
-
-      expect(context.homeStay, homeStayTruth);
-      expect(context.routineIndex, -1.0);
-      expect(context.stops.length, 6);
-      expect(context.moves.length, 5);
-      expect(context.places.length, 4);
-    });
-
-    test('Features: Multiple days, multiple locations', () async {
-      /// Clean file every time test is run
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-      mf.usePriorContexts = true;
-
-      for (int i = 0; i < 5; i++) {
-        DateTime date = jan01.add(Duration(days: i));
-
-        /// Todays data
-        List<LocationSample> locationSamples = [
-          // 5 hours spent at home
-          LocationSample(pos1, date.add(Duration(hours: 0, minutes: 0))),
-          LocationSample(pos1, date.add(Duration(hours: 6, minutes: 0))),
-
-          LocationSample(pos2, date.add(Duration(hours: 8, minutes: 0))),
-          LocationSample(pos2, date.add(Duration(hours: 9, minutes: 0))),
-        ];
-
-        await mf.saveSamples(locationSamples);
-
-        /// Calculate and save context
-        MobilityContext context = await mf.computeFeatures(date: date);
-
-        double routineIndex = context.routineIndex;
-        double homeStay = context.homeStay;
-
-        expect(context.stops.length, 2);
-        expect(context.places.length, 2);
-        expect(context.moves.length, 1);
-
-        expect(homeStay, 6 / 9);
-
-        // The first day the routine index should be -1,
-        // otherwise 1 since the days are exactly the same
-        if (i == 0) {
-          expect(routineIndex, -1);
-        } else {
-          expect(routineIndex, 1);
-        }
-      }
-    });
-
-    test('Stops: Multiple days, multiple locations, with overlap', () async {
-      /// Clean file every time test is run
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-
-      for (int i = 0; i < 5; i++) {
-        DateTime date = jan01.add(Duration(days: i));
-
-        /// Todays data
-        List<LocationSample> locationSamples = [
-          // 5 hours spent at home
-          LocationSample(pos1, date.add(Duration(hours: 0, minutes: 0))),
-          LocationSample(pos1, date.add(Duration(hours: 6, minutes: 0))),
-
-          LocationSample(pos2, date.add(Duration(hours: 8, minutes: 0))),
-          LocationSample(pos2, date.add(Duration(hours: 9, minutes: 0))),
-
-          LocationSample(pos1, date.add(Duration(hours: 21, minutes: 0))),
-          LocationSample(
-              pos1, date.add(Duration(hours: 23, minutes: 59, seconds: 59))),
-        ];
-
-        await mf.saveSamples(locationSamples);
-
-        /// Calculate and save context
-        MobilityContext context = await mf.computeFeatures(date: date);
-
-        /// Verify that stops are not shared among days
-        /// This should not be the case since samples from
-        /// previous days are filtered out.
-        expect(context.stops.length, 3);
-        expect(context.places.length, 2);
-        expect(context.moves.length, 2);
-      }
-    });
-
     test('Test that samples persist', () async {
       /// Clean file every time test is run
-      await flushFiles();
+      flushFiles();
       MobilityFactory mf = MobilityFactory.instance;
 
       List<LocationSample> samples = [
@@ -273,17 +119,13 @@ void main() async {
       ];
 
       await mf.saveSamples(samples);
-
-      /// Calculate and save context
-      MobilityContext context = await mf.computeFeatures(date: jan01);
-
       final loaded = await mf.loadSamples();
-      printList(loaded);
+      expect(loaded.length, samples.length);
     });
 
     test('Remove duplicate samples', () async {
       /// Clean file every time test is run
-      await flushFiles();
+      flushFiles();
 
       List<LocationSample> samplesNoDuplicates = [
         // 5 hours spent at home
@@ -321,34 +163,8 @@ void main() async {
             pos1, jan01.add(Duration(hours: 23, minutes: 59, seconds: 59))),
       ];
 
-      printList(samplesWithDuplicates.map((e) => e.datetime).toSet().toList());
-      await MobilityFactory.instance.saveSamples(samplesNoDuplicates);
-
-      /// Calculate and save context
-      MobilityContext context1 =
-          await MobilityFactory.instance.computeFeatures(date: jan01);
-
-      /// Verify that stops are not shared among days
-      /// This should not be the case since samples from
-      /// previous days are filtered out.
-      expect(context1.stops.length, 3);
-      expect(context1.places.length, 2);
-      expect(context1.moves.length, 2);
-
-      flushFiles();
-
-      await MobilityFactory.instance.saveSamples(samplesWithDuplicates);
-
-      /// Calculate and save context
-      MobilityContext context2 =
-          await MobilityFactory.instance.computeFeatures(date: jan01);
-
-      /// Verify that stops are not shared among days
-      /// This should not be the case since samples from
-      /// previous days are filtered out.
-      expect(context2.stops.length, 3);
-      expect(context2.places.length, 2);
-      expect(context2.moves.length, 2);
+      final samples = MobilityFactory.uniqueElements(samplesWithDuplicates);
+      expect(samples.length, samplesNoDuplicates.length);
     });
 
     test('Save samples one at a time', () async {
@@ -377,243 +193,50 @@ void main() async {
     });
 
     test('Stream LocationSamples one by one', () async {
+      void onContext(MobilityContext mc) {
+        print(mc.toJson());
+      }
+
       flushFiles();
       DateTime date = jan01;
 
       List<LocationSample> samples = [
-        // 5 hours spent at home
+        /// Home
         LocationSample(pos1, date.add(Duration(hours: 0, minutes: 0))),
         LocationSample(pos1, date.add(Duration(hours: 6, minutes: 0))),
 
+        /// Out
         LocationSample(pos2, date.add(Duration(hours: 8, minutes: 0))),
-        LocationSample(pos2, date.add(Duration(hours: 9, minutes: 0))),
+        LocationSample(pos2, date.add(Duration(hours: 12, minutes: 0))),
 
-        LocationSample(pos1, date.add(Duration(hours: 21, minutes: 0))),
-        LocationSample(
-            pos1, date.add(Duration(hours: 23, minutes: 59, seconds: 59))),
+        /// Home
+        LocationSample(pos1, date.add(Duration(hours: 17, minutes: 0))),
+        LocationSample(pos1, date.add(Duration(hours: 22, minutes: 0))),
+
+        /// Home, New day
+        LocationSample(pos1, date.add(Duration(days: 1, hours: 0, minutes: 2))),
       ];
 
-      MobilityFactory mf = MobilityFactory.instance;
-      mf.saveEvery = 1;
+      /// Create stream controller to stream the individual samples
+      /// to the MobilityFactory instance
+      StreamController<LocationSample> controller =
+          StreamController.broadcast();
 
-      Stream<LocationSample> stream = Stream.fromIterable(samples);
+      /// Set up stream
+      MobilityFactory mf = MobilityFactory.instance;
+      await mf.startListening(controller.stream);
+
+      int expectedContexts = 3;
+
+      /// Listen to the Context stream
+      Stream<MobilityContext> contextStream = mf.contextStream;
+      contextStream.listen(expectAsync1(onContext, count: expectedContexts));
 
       // Stream all the samples one by one
-      await mf.startListening(stream);
-
-      // After all data has been streamed, load the saved data
-      final loaded = await mf.loadSamples();
-
-      // Verify that all data was streamed and stored
-      expect(loaded.length, samples.length);
-    });
-
-    test('Stream Location DTOs one by one', () async {
-      flushFiles();
-
-      // Array of location DTOs, mocking DTOs from an external plugin
-      List<LocationDTO> dtos = [
-        LocationDTO(111.123, 123.123),
-        LocationDTO(222.123, 123.123),
-        LocationDTO(333.123, 123.123),
-      ];
-
-      // Convert DTO array to DTO stream
-      Stream<LocationDTO> locationStream = Stream.fromIterable(dtos);
-
-      // Convert DTO stream to LocationSample stream via a mapper function
-      Stream<LocationSample> stream = locationStream.map((dto) =>
-          LocationSample(GeoLocation(dto.lat, dto.lon), DateTime.now()));
-
-      // Instantiate the MobilityFactory
-      MobilityFactory mf = MobilityFactory.instance;
-      mf.saveEvery = 1;
-
-      // Stream all the samples one by one
-      await mf.startListening(stream);
-
-      // After all data has been streamed, load the saved data
-      final loaded = await mf.loadSamples();
-
-      // Verify that all data was streamed and stored
-      expect(loaded.length, dtos.length);
-    });
-
-    test('Test for isolate', () async {
-      /// Clean file every time test is run
-      await flushFiles();
-      MobilityFactory mf = MobilityFactory.instance;
-
-      List<LocationSample> samples = [
-        // 5 hours spent at home
-        LocationSample(pos1, jan01.add(Duration(hours: 0, minutes: 0))),
-        LocationSample(pos1, jan01.add(Duration(hours: 6, minutes: 0))),
-
-        LocationSample(pos2, jan01.add(Duration(hours: 8, minutes: 0))),
-        LocationSample(pos2, jan01.add(Duration(hours: 9, minutes: 0))),
-
-        LocationSample(pos1, jan01.add(Duration(hours: 21, minutes: 0))),
-        LocationSample(
-            pos1, jan01.add(Duration(hours: 23, minutes: 59, seconds: 59))),
-      ];
-
-      await mf.saveSamples(samples);
-
-      /// Calculate and save context
-      MobilityContext context = await mf.computeFeatures(date: jan01);
-
-      final loaded = await mf.loadSamples();
-      printList(loaded);
-    });
-
-    test('Multiple days, stream', () async {
-      /// Clean file every time test is run
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-      List<LocationSample> data = [];
-      mf.saveEvery = 1;
-
-      for (int i = 0; i < 5; i++) {
-        DateTime date = jan01.add(Duration(days: i));
-
-        /// Todays data
-        List<LocationSample> locationSamples = [
-          // 5 hours spent at home
-          LocationSample(pos1, date.add(Duration(hours: 0, minutes: 0))),
-          LocationSample(pos1, date.add(Duration(hours: 6, minutes: 0))),
-
-          LocationSample(pos2, date.add(Duration(hours: 8, minutes: 0))),
-          LocationSample(pos2, date.add(Duration(hours: 9, minutes: 0))),
-
-          LocationSample(pos1, date.add(Duration(hours: 21, minutes: 0))),
-          LocationSample(
-              pos1, date.add(Duration(hours: 23, minutes: 59, seconds: 59))),
-        ];
-
-        data.addAll(locationSamples);
+      for (LocationSample s in samples) {
+        controller.add(s);
       }
-
-      Stream<LocationSample> stream = Stream.fromIterable(data);
-      mf.startListening(stream);
-
-      List<DateTime> oddOrderedDates = [
-        DateTime(2020, 01, 04),
-        DateTime(2020, 01, 05),
-        DateTime(2020, 01, 04),
-        DateTime(2020, 01, 03),
-        DateTime(2020, 01, 01),
-        DateTime(2020, 01, 02)
-      ];
-      for (var date in oddOrderedDates) {
-        MobilityContext context = await mf.computeFeatures(date: date);
-        expect(context.stops.length, 3);
-        expect(context.places.length, 2);
-        expect(context.moves.length, 2);
-      }
-    });
-
-    test('Multiple days, nested', () async {
-      /// Clean file every time test is run
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-      mf.saveEvery = 1;
-
-      List<DateTime> dates = [
-        DateTime(2020, 01, 01),
-        DateTime(2020, 01, 02),
-        DateTime(2020, 01, 03),
-        DateTime(2020, 01, 04),
-        DateTime(2020, 01, 05),
-      ];
-
-      int n = dates.length;
-
-      for (int i = 0; i < n; i++) {
-        DateTime _date = dates[i];
-
-        /// Todays data
-        List<LocationSample> locationSamples = [
-          // 5 hours spent at home
-          LocationSample(pos1, _date.add(Duration(hours: 0, minutes: 0))),
-          LocationSample(pos1, _date.add(Duration(hours: 6, minutes: 0))),
-
-          LocationSample(pos2, _date.add(Duration(hours: 8, minutes: 0))),
-          LocationSample(pos2, _date.add(Duration(hours: 9, minutes: 0))),
-
-          LocationSample(pos1, _date.add(Duration(hours: 21, minutes: 0))),
-          LocationSample(
-              pos1, _date.add(Duration(hours: 23, minutes: 59, seconds: 59))),
-        ];
-
-        Stream<LocationSample> stream = Stream.fromIterable(locationSamples);
-        mf.startListening(stream);
-
-        for (var date in dates.sublist(0, i + 1)) {
-          MobilityContext context = await mf.computeFeatures(date: date);
-          expect(context.stops.length, 3);
-          expect(context.places.length, 2);
-          expect(context.moves.length, 2);
-        }
-        mf.stopListening();
-      }
-    });
-
-    test('Custom set function', () async {
-      /// Clean file every time test is run
-      flushFiles();
-
-      MobilityFactory mf = MobilityFactory.instance;
-      mf.saveEvery = 1;
-
-      List<LocationSample> data = [];
-
-      List<DateTime> dates = [
-        DateTime(2020, 01, 01),
-        DateTime(2020, 01, 02),
-        DateTime(2020, 01, 03),
-        DateTime(2020, 01, 04),
-        DateTime(2020, 01, 05),
-      ];
-
-      int n = dates.length;
-
-      for (int i = 0; i < n; i++) {
-        DateTime _date = dates[i];
-
-        /// Todays data
-        List<LocationSample> locationSamples = [
-          // 5 hours spent at home
-          LocationSample(pos1, _date.add(Duration(hours: 0, minutes: 0))),
-          LocationSample(pos1, _date.add(Duration(hours: 6, minutes: 0))),
-
-          LocationSample(pos2, _date.add(Duration(hours: 8, minutes: 0))),
-          LocationSample(pos2, _date.add(Duration(hours: 9, minutes: 0))),
-
-          LocationSample(pos1, _date.add(Duration(hours: 21, minutes: 0))),
-          LocationSample(
-              pos1, _date.add(Duration(hours: 23, minutes: 59, seconds: 59))),
-        ];
-        data.addAll(locationSamples);
-      }
-
-      Stream<LocationSample> stream = Stream.fromIterable(data);
-      mf.startListening(stream);
-
-      List<Stop> stops = [];
-
-      for (var date in dates) {
-        MobilityContext context = await mf.computeFeatures(date: date);
-        stops.addAll(context.stops);
-        stops.addAll(context.stops);
-        stops.addAll(context.stops);
-      }
-
-      stops.shuffle();
-      final stopSet = MobilityFactory.timestampSet(stops);
-      expect(stopSet.length, 15);
-      printList(stopSet);
+      controller.close();
     });
   });
 }